requirements.txt
aiohttp==3.8
aiosignal==1.2
async-timeout==4.0
asyncio==3.4
attrs==21.4
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1
frozenlist==1.3
idna==3
multidict==6.0
numpy==1.23
pandas==1.4
python-dateutil==2.8
pytz==2022
six==1.16
yarl==1.7
